moving model classes to .api subpackage
