+++ app/common/LogicalGroup.scala
@@ -17,6 +17,7 @@
 package common
 
 import models._
+import models.api.{VatFinancials, VatSicAndCompliance, VatTradingDetails}
 
 trait LogicalGroup[T] {
