+++ app/models/api/TradingName.scala
@@ -14,21 +14,14 @@
  * limitations under the License.
  */
 
-package models
+package models.api
 
-import java.time.LocalDate
+import play.api.libs.json.Json
 
-import play.api.libs.functional.syntax._
-import play.api.libs.json._
+case class TradingName(selection: Boolean, tradingName: Option[String])
 
+object TradingName {
 
-case class VatStartDate(selection: String, startDate: Option[LocalDate])
-
-object VatStartDate extends VatStartDateValidator {
-
-  implicit val format: OFormat[VatStartDate] = (
-    (__ \ "selection").format[String](vatStartDateValidator) and
-      (__ \ "startDate").formatNullable[LocalDate]
-    ) (VatStartDate.apply, unlift(VatStartDate.unapply))
+  implicit val format = Json.format[TradingName]
 
 }
