+++ app/models/api/Validation.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package models
+package models.api
 
 import play.api.libs.json.Reads.pattern
 import play.api.libs.json._
