+++ app/models/api/VatAccountingPeriod.scala
@@ -14,17 +14,21 @@
  * limitations under the License.
  */
 
-package models
+package models.api
 
+import play.api.libs.functional.syntax._
 import play.api.libs.json._
 
-case class VatTradingDetails(
-                              vatChoice: VatChoice,
-                              tradingName: TradingName
-                            )
+case class VatAccountingPeriod(
+                                frequency: String, // "monthly" or "quarterly"
+                                periodStart: Option[String] = None // "jan_apr_jul_oct", "feb_may_aug_nov" or "mar_jun_sep_dec"
+                              )
 
-object VatTradingDetails {
+object VatAccountingPeriod extends VatAccountingPeriodValidator {
 
-  implicit val format: OFormat[VatTradingDetails] = Json.format[VatTradingDetails]
+  implicit val format = (
+    (__ \ "frequency").format[String](frequencyValidator) and
+      (__ \ "periodStart").formatNullable[String](periodStartValidator)
+    ) (VatAccountingPeriod.apply, unlift(VatAccountingPeriod.unapply))
 
 }
