+++ app/models/api/VatFinancials.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package models
+package models.api
 
 import play.api.libs.functional.syntax._
 import play.api.libs.json._
