+++ app/models/api/VatTradingDetails.scala
@@ -14,21 +14,17 @@
  * limitations under the License.
  */
 
-package models
+package models.api
 
-import play.api.libs.functional.syntax._
 import play.api.libs.json._
 
-case class VatAccountingPeriod(
-                                frequency: String, // "monthly" or "quarterly"
-                                periodStart: Option[String] = None // "jan_apr_jul_oct", "feb_may_aug_nov" or "mar_jun_sep_dec"
-                              )
+case class VatTradingDetails(
+                              vatChoice: VatChoice,
+                              tradingName: TradingName
+                            )
 
-object VatAccountingPeriod extends VatAccountingPeriodValidator {
+object VatTradingDetails {
 
-  implicit val format = (
-    (__ \ "frequency").format[String](frequencyValidator) and
-      (__ \ "periodStart").formatNullable[String](periodStartValidator)
-    ) (VatAccountingPeriod.apply, unlift(VatAccountingPeriod.unapply))
+  implicit val format: OFormat[VatTradingDetails] = Json.format[VatTradingDetails]
 
 }
